Update README.md
This project is Food Delivery App. There you can register or login and see the reccomendation of meals . Then you can do an order and get it delivered. Moreover there is the function in this app that allows you to do payment through this app and see delivery process progres.